% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{bulk.dat}
\alias{bulk.dat}
\title{bulk.dat}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1572 rows and 126 columns.
}
\usage{
bulk.dat
}
\description{
Simulated bulk tissue data of 126 samples. It's part of the simulated data used in HiDecon.
}
\keyword{datasets}
