% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiDecon.R
\name{Check.KKT}
\alias{Check.KKT}
\title{Check.KKT.}
\usage{
Check.KKT(grad, pi, tol = 1e-06)
}
\arguments{
\item{grad}{gradient of the objective function.}

\item{pi}{the update of cellular fraction estimates.}

\item{tol}{tolerance for the KKT conditions, default is 1e-6.}
}
\value{
logic.If KKT conditions satisfied, TRUE. Otherwise, FALSE.
}
\description{
Check the KKT conditions as described in the HiDecon paper.
}
