% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiDecon.R
\name{Create.A}
\alias{Create.A}
\title{Create.A.}
\usage{
Create.A(Signature.list, Cell.size.list = NULL, standardize = T)
}
\arguments{
\item{Signature.list}{a length L list of #markers_l * #cell_types signature matrices for the L layers.}

\item{Cell.size.list}{a length L list of cell size vectors. Each vector is of length K.l=#cell_types in layer l.}

\item{standardize}{if the loss should be standardized by the number of marker genes. Default is TRUE.}
}
\value{
A list to be used in function Est.AllPi.
}
\description{
Create the A.tilde in the HiDecon paper.
}
