% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiDecon.R
\name{Create.B}
\alias{Create.B}
\title{Create.B.}
\usage{
Create.B(B.list, standardize = T)
}
\arguments{
\item{B.list}{list: contains cell type mapping matrices (from B1,2).}

\item{standardize}{if the loss should be standardized by the number of cell types. Default is true.}
}
\value{
A matrix to be used in function Est.AllPi.
}
\description{
Create the B.tilde in the HiDecon paper.
}
