% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiDecon.R
\name{Est.AllPi}
\alias{Est.AllPi}
\title{Est.AllPi}
\usage{
Est.AllPi(Y.list, A, B, lambda, Pi.start = NULL, max.iter = 10000, tol = 1e-06)
}
\arguments{
\item{Y.list}{a length L list of #individuals x #markers bulk data matrices. The dimension of the lth element of this list is #individuals x #markers_l.}

\item{A}{the output from Create.A.}

\item{B}{the output from Create.B.}

\item{lambda}{numerical, the penalty parameter lambda in the HiDecon paper.}

\item{Pi.start}{vector, non-negative starting point for the estimation of sample i. Default is NULL, then function will assign the starting point.}

\item{max.iter}{the maximum number of iterations, default is 1e4.}

\item{tol}{tolerance for the KKT conditions, default is 1e-6.}
}
\value{
a list with elements:
   * Pi: a #individuals x K.L matrix of cellular fractions
   * n.iter: number of iterations for each individual
   * out: whether the optimization converged. 0:yes, 1:no
}
\description{
Estimate cellular fractions for all samples.
}
