% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiDecon.R
\name{EstPi}
\alias{EstPi}
\title{EstPi.}
\usage{
EstPi(H, H.inv, Aty, K.L, pi.start = NULL, max.iter = 10000, tol = 1e-06)
}
\arguments{
\item{H}{matrix, Hessian matrix in the HiDecon paper.}

\item{H.inv}{inverse of H.}

\item{Aty}{the ith row of matrix Y*A.}

\item{K.L}{number of cell types in layer L.}

\item{pi.start}{vector, non-negative starting point for the estimation of sample i. Default is NULL, then function will assign the starting point.}

\item{max.iter}{the maximum number of iterations, default is 1e4.}

\item{tol}{tolerance for the KKT conditions, default is 1e-6.}
}
\value{
a list with elements:
   * Pi: cellular fractions for one sample
   * n.iter: number of iterations to convergence
   * out: whether the optimization converged. 0:yes, 1:no
}
\description{
Estimate cellular fractions for one sample.
}
