% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiDecon.R
\name{GeneratePseudoBulk}
\alias{GeneratePseudoBulk}
\title{GeneratePseudoBulk.}
\usage{
GeneratePseudoBulk(
  bulk,
  ref,
  cell_type,
  type_order,
  test = "wilcox",
  nmrk = 50,
  seed = 123
)
}
\arguments{
\item{bulk}{matrix, bulk data of genes by samples. It should be normalized.}

\item{ref}{matrix: Single cell data matrix (genes by cells) after normalization.}

\item{cell_type}{vector: cell type label of single cell reference.}

\item{type_order}{vector: specify cell type order in the bottom layer of tree.}

\item{test}{the "test.type" in scran::findMarkers. The test used to find markers. test  = c("t", "wilcox", "binom"). Default is "wilcox".}

\item{nmrk}{numerical, number of marker genes chosen for each cell type, default is 50.}

\item{seed}{random seed for resampling. Default is 123.}
}
\value{
a list with elements:
   * bulk.sim: bulk data simulated using fractions estimated by NNLS
   * frac.sim: fractions estimated by NNLS
}
\description{
Generate bulk data surrogate for tuning parameter selection by resampling from single cell reference based on NNLS fraction estimates.
}
