% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiDecon.R
\name{GeneratesigReference}
\alias{GeneratesigReference}
\title{GeneratesigReference.}
\usage{
GeneratesigReference(ref, cell.types)
}
\arguments{
\item{ref}{matrix (genes * cells): single cell reference.}

\item{cell.types}{vector: cell type label of reference.}
}
\value{
matrix: signature matrix
}
\description{
General signature matrix calculator.
}
