% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiDecon.R
\name{HiDecon}
\alias{HiDecon}
\title{HiDecon.}
\usage{
HiDecon(
  bulk,
  ref,
  B,
  cell_type,
  type_order,
  lambda = 40,
  Pi.start = NULL,
  max.iter = 10000,
  tol = 1e-06,
  test = "wilcox",
  nmrk = 50
)
}
\arguments{
\item{bulk}{matrix, bulk data of genes by samples. It should be count per million normalized.}

\item{ref}{matrix: Single cell data matrix (genes by cells) after count per million normalization.}

\item{B}{list: contains cell type mapping matrices (same length as tree depth).}

\item{cell_type}{vector: cell type label of single cell reference.}

\item{type_order}{vector: specify cell type order in the bottom layer of tree.}

\item{lambda}{numerical, the penalty parameter lambda in the HiDecon paper.}

\item{Pi.start}{vector, non-negative starting point for the estimation of sample i. Default is NULL, then function will assign the starting point.}

\item{max.iter}{the maximum number of iterations, default is 1e4.}

\item{tol}{tolerance for the KKT conditions, default is 1e-6.}

\item{test}{the "test.type" in scran::findMarkers. The test used to find markers. test  = c("t", "wilcox", "binom"). Default is "wilcox".}

\item{nmrk}{numerical, number of marker genes chosen for each cell type, default is 50.}
}
\value{
matrix, HiDecon cellular fraction estimates for samples.
}
\description{
Estimate cellular fractions for bulk tissue data using single cell reference with hierarchical tree.
}
