% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiDecon.R
\name{HiDecon_input}
\alias{HiDecon_input}
\title{HiDecon_input.}
\usage{
HiDecon_input(bulk, ref, cell_type, B, type_order, test = "wilcox", nmrk = 50)
}
\arguments{
\item{bulk}{matrix, bulk data of genes by samples. It should be count per million normalized.}

\item{ref}{matrix: Single cell data matrix (genes by cells) after count per million normalization.}

\item{cell_type}{vector: cell type label of single cell reference.}

\item{B}{list: contains cell type mapping matrices (same length as tree depth).}

\item{type_order}{vector: specify cell type order in the bottom layer of tree.}

\item{test}{the "test.type" in scran::findMarkers. The test used to find markers. test  = c("t", "wilcox", "binom"). Default is "wilcox".}

\item{nmrk}{numerical, number of marker genes chosen for each cell type, default is 50.}
}
\value{
a list with input elements for function HiDecon and select_HiDecon.
}
\description{
Calculate inputs for HiDecon and
}
