% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiDecon.R
\name{HiDecon_marker}
\alias{HiDecon_marker}
\title{HiDecon_marker}
\usage{
HiDecon_marker(ref, cell_type, B, type_order, test = "wilcox", nmrk = 50)
}
\arguments{
\item{ref}{matrix: Single cell data matrix (cells * genes) after count per million normalization.}

\item{cell_type}{vector: cell type label of single cell reference.}

\item{B}{list: contains cell type mapping matrices (same length as tree depth).}

\item{type_order}{vector: specify cell type order in the bottom layer of tree.}

\item{test}{the "test.type" in scran::findMarkers. The test used to find markers. test  = c("t", "wilcox", "binom"). Default is "wilcox".}

\item{nmrk}{numerical, number of marker genes chosen for each cell type, default is 50.}
}
\value{
a list in which elements are marker gene names for each layer of tree.
}
\description{
Select marker genes for each layer of the hierarchical tree.
}
