% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiDecon.R
\name{my.CCC}
\alias{my.CCC}
\title{my.CCC.}
\usage{
my.CCC(est, tr)
}
\arguments{
\item{est}{matrix, estimated cellular fractions (samples * cell types)}

\item{tr}{matrix, ground truth (sample * cell types)}
}
\value{
CCC and mean CCC
}
\description{
Calculate Lin's concordance correlation coefficients between estimates and ground truth for cell types and the mean CCC.
}
