% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{order.type}
\alias{order.type}
\title{order.type}
\format{
An object of class \code{character} of length 10.
}
\usage{
order.type
}
\description{
Cell type order of the example data bottome layer of the hierarchical tree.
}
\keyword{datasets}
