% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ref.dat}
\alias{ref.dat}
\title{ref.dat}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1572 rows and 100 columns.
}
\usage{
ref.dat
}
\description{
The single cell reference used in the HiDecon example data. It's part of the simulated data used in HiDecon.
}
\keyword{datasets}
