% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ref.type}
\alias{ref.type}
\title{ref.type}
\format{
An object of class \code{character} of length 100.
}
\usage{
ref.type
}
\description{
Cell type labels of the single cell data ref.dat.
}
\keyword{datasets}
