% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiDecon.R
\name{reference}
\alias{reference}
\title{reference.}
\usage{
reference(ref, ref.type, B, type_order, layer)
}
\arguments{
\item{ref}{matrix: Single cell data matrix (cells * genes) after count per million normalization and log2 fold change.}

\item{ref.type}{vector: cell type label of reference.}

\item{B}{list: contains cell type mapping matrices (same length as tree depth).}

\item{type_order}{vector: specify cell type order in the bottom layer of tree.}

\item{layer}{integer: the layer needed to be calculated.}
}
\value{
matrix: signature matrix for the specified layer
}
\description{
Calculate signature matrix (reference) for some layer.
}
